+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
  *Milestone: Easily understandable runtime errors*
  *   Allow breakpoints and stepping to work together in debug
+ *   API_uploadCode() allows to immediately run the code: Regarding preload it defined whether the code should be preloaded in the whiteboard or run when the user clicks run
+ *   Breakpoints in debug now count how many times the app has gone through them (most useful when the breakpoint exists but is disabled)
 
 ## 2.1 (2015-10-06)